requirements.txt
asgiref==3.10.0
asttokens==3.0.0
black==25.11.0
boto3==1.41.3
botocore==1.41.3
cfgv==3.4.0
click==8.3.0
colorama==0.4.6
comm==0.2.3
debugpy==1.8.17
decorator==5.2.1
distlib==0.4.0
Django==5.2.7
django-extensions==4.1
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djhtml==3.0.10
executing==2.2.1
filelock==3.20.0
identify==2.6.15
ipykernel==7.0.1
ipython==9.6.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jmespath==1.0.1
jupyter_client==8.6.3
jupyter_core==5.9.1
matplotlib-inline==0.2.1
mypy_extensions==1.1.0
nest-asyncio==1.6.0
nodeenv==1.9.1
packaging==25.0
parso==0.8.5
pathspec==0.12.1
pillow==12.0.0
platformdirs==4.5.0
pre_commit==4.4.0
prompt_toolkit==3.0.52
psutil==7.1.1
psycopg2-binary==2.9.11
pure_eval==0.2.3
Pygments==2.19.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytokens==0.3.0
PyYAML==6.0.3
pyzmq==27.1.0
s3transfer==0.15.0
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
tornado==6.5.2
traitlets==5.14.3
tzdata==2025.2
urllib3==2.5.0
virtualenv==20.35.4
wcwidth==0.2.14
whitenoise==6.11.0